Rearrange the .h files in SimCalorimetry/HcalZeroSuppressionProducers
