+++ SimCalorimetry/HcalZeroSuppressionProducers/interface/HcalZSAlgoRealistic.h
@@ -1,7 +1,7 @@
-#ifndef SIMCALORIMETRY_HCALZEROSUPPRESSIONALGOS_HCALZSALGOREALISTIC_H
-#define SIMCALORIMETRY_HCALZEROSUPPRESSIONALGOS_HCALZSALGOREALISTIC_H 1
+#ifndef SIMCALORIMETRY_HCALZEROSUPPRESSIONPRODUCERS_HCALZSALGOREALISTIC_H
+#define SIMCALORIMETRY_HCALZEROSUPPRESSIONPRODUCERS_HCALZSALGOREALISTIC_H 1
 
-#include "HcalZeroSuppressionAlgo.h"
+#include "SimCalorimetry/HcalZeroSuppressionProducers/interface/HcalZeroSuppressionAlgo.h"
 
 /** \class HcalZSAlgoRealistic
  *
