--- SimCalorimetry/HcalZeroSuppressionProducers/src/SealModule.cc
@@ -1,5 +1,0 @@
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-#include "HcalRealisticZS.h"
-
-DEFINE_FWK_MODULE(HcalRealisticZS);
